refactor: move icon components in icons folder
